Move LatentsToImageInvocation to its own file. No functional changes.
